grpc-transformer/requirements.txt
grpcio-tools
transformers
